submissions/50-honeycomb-distributed-tracing/requirements.txt
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-http>=1.22.0
python-dotenv>=1.0.0
